Add better default return value for scalar exec
Changes `execFirst` and `execScalar` so that they no longer raise an
error if the result set is empty. Instead `execFirst` returns an
`Option[T]`, and `execScalar` returns an empty `OdbcValue`.

Reason for not using an empty `OdbcRowSet` for `execFirst` is that it
makes `OdbcRowSet` less predictable, because the use case is to index
it, and an empty `OdbcRowSet` will always raise an index out of bounds
error in this case.