Add UTF-8 encoding for `varchar` columns
Previously `varchar` columns were read literally, so the encoding was
unknown, and the developer had to manage this.

Now use UTF-16 API to make the driver/DBMS convert the column, so that
it can be converted to UTF-8 in this library.

Fix these kinds of tests to reflect the changes:

* How data is encoded in database storage
* Use of `otCharArray` for strings is mostly replaced with `otWideArray`