bindParam: Add proper empty string binding
This improves how empty strings are bound to parameters in queries. The
fix in `47ca935` did not address all problems. Specifically, binding
parameters to NULL if they're empty strings is problematic for some
drivers (ODBC Driver 17 for SQL Server in testing).

This changes so that, if the string is empty, bind to a static array
containing zeroes. It is required by SQLBindParameter that the "buffer"
is not nil. Because we cannot get the address to the buffer in empty Nim
containers, a static buffer solves this issue.

---

Also changed "SQL data type" of array types to use "*VAR*" variant
instead.

Previously when an empty string is bound to a parameter, and the query
is executed, a string with a space (' ') is given to SQL Server. It is
expected that a string without space is given to SQL Server. Changing
"SQL data type" as described fixes this. I'm unsure why.