widestrx: Add support for `openArray[char]` in `utf8To16`
This adds support for the `openArray` view of `string` to be given as
argument to `utf8To16`. This just extends the functionality. With this,
`seq[char]` can be used as argument as well.

For this to work, some additional routines had to be copied from
`system/widestrs`, modified to take an `openArray[char]` parameter.