Drop `withTran` template
This causes some issues. The following code is an issue:

```nim
conn.withTran:
  if somecond():
    return
  doThis()
```

In this case the transaction will not be committed or rollbacked on the
`return` statement. The template doesn't have any finally block. This
may be misleading to the user, so explicit transactions is better style
for maintenance purposes.

Users may implement `withTran` themselves if they understand the
problems. Regardless it's important not to recommend `withTran` in the
public API.