Fix ODBC buffer overread on query parameters
Issue is that the ODBC driver "ODBC Driver for SQL Server 17" and 18
(and maybe older) can do a buffer overread of string parameters in
queries. ODBC returns an error saying "String data, right truncation".
This does not always occur. It happens because the driver assumes that
the string is zero-terminated, which odbcn does not guarantee. This
driver assumes that because the "StrLen_or_IndPtr" argument is `nil`. It
does not use "ColumnSize" or "BufferLength" parameters like the FreeTDS
driver does. Instead the value pointed to by "StrLen_or_IndPtr" is
expected to hold the buffer size.

This fixes the buffer overread by setting "StrLen_or_IndPtr" argument to
the buffer size. The parameters "BufferSize", "ColumnSize" and
"StrLen_or_IndPtr" to SQLBindParameter for array types are redundant
with each other. Minimized redundancy by not setting "ColumnSize". Added
FAQ to explain this.

Can now stop having a global buffer for empty arrays, because the buffer
size was previously incorrectly specified. Now it works for empty
arrays.

This error is tested by running odbc.nim integration tests with "ODBC
Driver 18 for SQL Server".