Drop ORM type generation; Rework exec-extensions
This adds several features. It was not separated into several commits
because it was most convenient this way for refactoring the tests.

* Drop type generation for prep-calls with connection string
* Drop `execScalar` and `execFirst` for `first` and `firstOrDefault`
  that work on the result set, instead of replacing `exec`
* Add `get(int): Option[OdbcValue]` for OdbcRowSet, which fascilitates
  use of `first(rs, OdbcValue)`

--- ORM type generation rationale ---

I never used this feature in practice. It was too inconvenient to have
to supply a connection string at compile-time. If this is still desired,
the `prep` macro can be overridden in another module or outside of this
package for special purposes.

--- Drop execScalar/execFirst ---

Motivation is to reduce redundancy in the API, and allow other return
types than OdbcRowSet. execScalar and execFirst are just wrappers of
`exec`, and there are multiple `exec`s defined in odbcn with different
signatures (e.g. prepared vs. non-prepared). As such this change makes
it separate from `exec` and just a method on the result set.

Upgrading:

* `conn.execScalar(qry)` -> `conn.exec(qry).firstOrDefault(OdbcValue)`
  * Keep in mind that you can use any type instead of `OdbcValue`, such
    as `int` or an object
* `conn.execFirst(qry)` -> `conn.exec(qry).first`

Drop routines for converting `OdbcValue` to other types. Instead prefer
that the user uses the typed interface, i.e. getting data into a type
other than `OdbcValue`/`OdbcRowSet`.