Add typed overload functionality
This properly makes it possible to overload `getData`, `bindParam` and
`bindCol`. While previously they had questionable support, it is now
easier to overload, better documented, and tested.

See `dataschema` module for documentation. It is intended that the user
uses the procs exported in that module to create their own overload.

Also simplified `bindParams` and `getDatas` code. Now they're able to
call `bindParam`/`getData` for an `object` type if the overload exists.
Previously the `object` would always be converted to a `getData`/...
call for each field.

Split error handling into `error` module so it can be used by other
modules without importing the `core` module.