Fix empty string causing null constraint error on SQLite
This library sends `nil` to the ODBC C library when binding the empty
string. This works for SQL Server. However, the SQLite ODBC driver
interprets the `nil` buffer as the SQL `NULL` value. This works fine in
most situations as NULL is converted to the empty string in SQLite.
However when binding a parameter that binds to a non-nullable column in
a table, the SQLite ODBC library complains about "null constraint
error". It is saying that it refuses to set a non-nullable column to
NULL.

Fixed by sending an empty cstring to the ODBC C library when binding
empty strings.